Split FieldDirectiveResolverInterface
